multi-hashing functionality added

+++ include/skabloom.h
@@ -1,18 +1,19 @@
 #include <stddef.h>
+#include <stdint.h>
 
 typedef struct skabloom_t {
-  size_t size;
-  size_t num_hashes;
-  size_t num_items;
+  uint64_t size;
+  int num_hashes;
+  uint64_t num_items;
   double max_error_rate;
   char *bitmap;
-  size_t num_filled;
+  uint64_t num_filled;
   double error_rate;
   struct skabloom_t *next;
 
 } skabloom_t;
 
-skabloom_t *skabloom_create(size_t value_hint, double max_error_rate);
+skabloom_t *skabloom_create(uint64_t value_hint, double max_error_rate);
 
 void skabloom_destroy(skabloom_t *s);
